requirements.txt
Flask
Flask-CORS
requests
Pillow
python-dotenv